Added Button Functionality
When I pulled this morning no buttons really had functionality, at least on my end and I think that was because of the way that the button class was setup. I have restructured the button class as well as the undo and change cursor functions so that they function correctly. Additionally there is room for an infinite amount of cursors with this new method. Lastly, functionality for saving and uploading drawings is fully implemented. Current bugs that I have seen and have been unable to remedy: Stray lines occurring when clicking buttons, these cannot be undo'd for some reason. The last bug, and it isnt really a bug, is that when you upload it doesn't clear the previous drawings on the page. Additionally there is no way to undo the upload. All solutions I could find for this sacrificed the transparent background.